Move analysis page to flaky
